+++ modpack/changelog.md
@@ -11,6 +11,11 @@ Replace these mods by their modrinth version:
 - CoroUtil
 - ParCool!
 
+Add:
+- Create
+- Flywheel
+- KubeJS Create
+
 ## 0.10.2
 
 **Theoretically compatible with 0.10.0 and 0.10.1 but not tested. Servers are highly encourage to upgrade**